ajout token and modif user query mutations server

+++ server/src/graphql/resolvers/mutations.ts
@@ -1,59 +1,43 @@
 import { prisma } from "../../database"
-import * as bcrypt from 'bcrypt';
+import * as bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+import { checkToken } from "../middlewares/resolversMiddlewares";
+
 
 export const Mutation = {
 
-  createProject: (parent: any, args:{ description: string, name: string, deadline: string , createdAt: string}) => {
-    // A REVOIR LOCAL DATE !!
-    // const dateInFr = new Date(args.deadline).toISOString();
-    // const dateInLocal= new Date(dateInFr).toLocaleString("fr-FR");
-    return prisma.projects.create({
-      data: {
-        description: args.description,
-        name: args.name,
-        deadline: args.deadline,
-        createdAt: args.createdAt
-      }
-    })
-  },
-  // createProjectNested: (parent: any, args:{id: string, description: string, name: string, deadline: string , createdAt: string, user_project: string}) => {
-  //   // A REVOIR LOCAL DATE !!
-  //   // const dateInFr = new Date(args.deadline).toISOString();
-  //   // const dateInLocal= new Date(dateInFr).toLocaleString("fr-FR");
-  //   return prisma.projects.create({
-  //     data: {
-  //       description: args.description,
-  //       name: args.name,
-  //       deadline: new Date(args.deadline),
-  //       createdAt: new Date(args.createdAt),
-  //       // join table
-  //       user_project: {
-  //         connect: [{id: Number(args.id)}]
-  //       }
-  //     }
-  //   })
-  // }
+  login: async (parent: any, args: { email: string; password: string }, context: any) => {
+    try {
+      const user = await prisma.users.findUnique({ where: { email: args.email } })
+      if (!user) throw new Error('No user with that email')
+      const isValidPassword = await bcrypt.compare(args.password, user.password)
+      if (!isValidPassword) throw new Error('Incorrect password')
 
-  updateProject: (parent: any, args:{ id: String, description: string, name: string, deadline: string}) => {
-    return prisma.projects.update({
-      where: { id: Number(args.id) },
-      data: {
-        description: args.description !== null ? args.description : undefined,
-        name: args.name !== null ? args.name : undefined,
-        deadline: args.deadline !== null ? args.deadline : undefined,
-      }
-    })
+      const token = jwt.sign(
+        {
+          id: user.id,
+          email: user.email,
+          role: user.roles
+        },
+        process.env.PRIVATE_KEY,
+        {
+          expiresIn: '1d',
+        }
+      )
+
+      return { token, user }
+    } catch (error) {
+      throw new Error(error.message)
+    }
   },
 
   createUser: async (parent: any, args:{firstname: string, lastname: string, email: string, password: string, tel: number})=>{
-
     const password: string= await cryptagePassword(args.password)
 
     async function cryptagePassword(password: string){
         const salt = await bcrypt.genSalt();
         const passwordHashed= await bcrypt.hash(password, salt);
         return passwordHashed;
-        
     }
     return prisma.users.create({
       data: {
@@ -67,19 +51,22 @@ export const Mutation = {
     })
   },
 
-  updateUser: async(parent: any, args:{id: String, firstname: string, lastname: string, email: string, password: string, tel: number})=>{
+  updateUser: async(parent: any, args:{id: String, firstname: string, lastname: string, email: string, password: string, tel: number}, decodedToken: any)=>{
+    checkToken(decodedToken)
+    const idUser= decodedToken.id;
+    const idUserToUpdate= Number(args.id)
+    const roleUser= decodedToken.role;
+    if(idUser !== idUserToUpdate && roleUser !== "ADMIN") throw new Error("UPDATE FORBIDDEN");
+
+    const user= await prisma.users.findUnique({where: {id: Number(args.id)}});
+    if(!user) throw new Error('id invalid')
   
-      console.log(args);
-      const user= await prisma.users.findUnique({where: {id: Number(args.id)}});
-      
-      if(!user){
-        throw new Error('id invalid')
-      }
-      const email= args.email? await checkEmail(args.email): null;
-      const firstname= args.firstname;
-      const lastname= args.lastname;
-      const tel= args.tel;
-      const password= args.password;
+
+    const email= args.email? await checkEmail(args.email): user.email;
+    const firstname= args.firstname? args.firstname: user.firstname;
+    const lastname= args.lastname? args.lastname : user.lastname;
+    const tel= args.tel? args.tel : user.tel;
+   
       
       async function checkEmail(emailToCheck: string){
         const email= await prisma.users.findUnique({where: {email : emailToCheck}})
@@ -92,14 +79,19 @@ export const Mutation = {
       data: {
         firstname,
         lastname,
-        password,
         email,
         tel
       }
     })
   },
 
-  deleteUser:(parent: any, args:{id: String})=>{
+  deleteUser:(parent: any, args:{id: String}, decodedToken: any)=>{
+    checkToken(decodedToken);
+    const idUser= decodedToken.id;
+    const idUserToDelete= Number(args.id);
+    const roleUser= decodedToken.role;
+    if(idUser !== idUserToDelete && roleUser !== "ADMIN") throw new Error("UPDATE FORBIDDEN");
+
     return prisma.users.delete({
       where:{
         id: Number(args.id)
@@ -107,6 +99,32 @@ export const Mutation = {
     })
   },
 
+
+  createProject: (parent: any, args:{ description: string, name: string, deadline: string , createdAt: string}, decodedToken: any) => {
+    checkToken(decodedToken);
+    // A REVOIR LOCAL DATE !!
+    // const dateInFr = new Date(args.deadline).toISOString();
+    // const dateInLocal= new Date(dateInFr).toLocaleString("fr-FR");
+    return prisma.projects.create({
+      data: {
+        description: args.description,
+        name: args.name,
+        deadline: args.deadline,
+        createdAt: args.createdAt
+      }
+    })
+  },
+  updateProject: (parent: any, args:{ id: String, description: string, name: string, deadline: string}) => {
+    
+    return prisma.projects.update({
+      where: { id: Number(args.id) },
+      data: {
+        description: args.description !== null ? args.description : undefined,
+        name: args.name !== null ? args.name : undefined,
+        deadline: args.deadline !== null ? args.deadline : undefined,
+      }
+    })
+  },
   deleteProject:(parent: any, args:{id: String})=>{
     //A FAIRE Supression en cascade
     return prisma.projects.delete({
@@ -114,5 +132,5 @@ export const Mutation = {
         id: Number(args.id)
       }
     })
-  }
+  },
 }
